Создан модуль user-service, тесты проходят
