+++ core/interaction-api/src/main/java/ru/practicum/exception/ConflictStateException.java
