+++ core/interaction-api/src/main/java/ru/practicum/exception/ConflictTimeException.java
