+++ core/interaction-api/src/main/java/ru/practicum/exception/ValidationException.java
