+++ core/interaction-api/src/main/java/ru/practicum/exception/handler/ErrorHandler.java
