+++ core/interaction-api/src/main/java/ru/practicum/exception/handler/ErrorResponse.java
