+++ core/interaction-api/src/main/java/ru/practicum/user/dto/UserDto.java
