+++ core/interaction-api/src/main/java/ru/practicum/user/dto/UserShortDto.java
