+++ core/pom.xml
@@ -14,6 +14,7 @@
     <modules>
         <module>main-service</module>
         <module>user-service</module>
+        <module>interaction-api</module>
     </modules>
 
     <properties>
