--- core/user-service/src/main/java/ru/practicum/UserServiceApp.java
@@ -1,15 +1,0 @@
-package ru.practicum;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-@ConfigurationPropertiesScan
-public class UserServiceApp {
-    public static void main(String[] args) {
-        SpringApplication.run(UserServiceApp.class, args);
-    }
-}
