+++ core/user-service/src/main/java/ru/practicum/user/controller/UserController.java
@@ -1,27 +1,27 @@
 package ru.practicum.user.controller;
 
+import jakarta.annotation.PostConstruct;
 import jakarta.validation.Valid;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.client.StatClient;
 import ru.practicum.user.dto.UserDto;
 import ru.practicum.user.dto.UsersDtoGetParam;
 import ru.practicum.user.service.UserService;
+
 
 import java.util.List;
 
 @Validated
 @RestController
-@AllArgsConstructor
+@RequiredArgsConstructor
 @RequestMapping("/admin/users")
 public class UserController {
     private static final String APP_NAME = "ewm-main-service";
 
     private final UserService userService;
-    private final StatClient statClient;
 
     @GetMapping
     public ResponseEntity<List<UserDto>> getUsers(@ModelAttribute @Valid UsersDtoGetParam usersDtoGetParam) {
@@ -45,4 +45,5 @@ public class UserController {
                 .status(HttpStatus.NO_CONTENT)
                 .body("Пользователь c id: " + userId + " удален");
     }
+
 }
