+++ core/user-service/src/main/java/ru/practicum/user/model/User.java
