+++ core/user-service/src/main/java/ru/practicum/user/repository/UserRepository.java
