+++ core/user-service/src/main/java/ru/practicum/user/service/UserService.java
