+++ core/user-service/src/main/resources/application.yaml
@@ -5,6 +5,7 @@ spring:
     import: "configserver:"
   cloud:
     config:
+
       fail-fast: true
       retry:
         use-random-policy: true
