+++ postman/ewm-main-service.json
@@ -1,7 +1,7 @@
 {
 	"info": {
 		"_postman_id": "b7d50c64-67b6-43c5-90ca-8e982296449a",
-		"name": "Test Explore With Me Plus - Main - Spring Cloud",
+		"name": "Test Explore With Me Plus - Main - Microservices",
 		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
 		"_exporter_id": "23073145",
 		"_collection_link": "https://universal-shadow-295426.postman.co/workspace/My-Workspace~4200f6aa-0504-44b1-8a1d-707d0dcbd5ce/collection/13708500-b7d50c64-67b6-43c5-90ca-8e982296449a?action=share&source=collection_link&creator=23073145"
