+++ postman/ewm-stat-service.json
@@ -1,7 +1,7 @@
 {
 	"info": {
 		"_postman_id": "fd4ffdc0-705a-450d-8d39-7c4eef7743f8",
-		"name": "Test Explore With Me Plus - Stat - Spring Cloud",
+		"name": "Test Explore With Me Plus - Stat - Microservices",
 		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
 		"_exporter_id": "23073145",
 		"_collection_link": "https://universal-shadow-295426.postman.co/workspace/My-Workspace~4200f6aa-0504-44b1-8a1d-707d0dcbd5ce/collection/13708500-fd4ffdc0-705a-450d-8d39-7c4eef7743f8?action=share&source=collection_link&creator=23073145"
